[metasrv] refactor: move trait impl for network to network.rs
