+++ metasrv/src/meta_service/log_entry.rs
@@ -12,14 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::convert::TryFrom;
-
 use async_raft::AppData;
 use serde::Deserialize;
 use serde::Serialize;
 
 use crate::meta_service::Cmd;
-use crate::meta_service::RaftMes;
 use crate::raft::types::RaftTxId;
 
 /// The application data request type which the `Metasrv` works with.
@@ -37,23 +34,3 @@ pub struct LogEntry {
 }
 
 impl AppData for LogEntry {}
-
-impl tonic::IntoRequest<RaftMes> for LogEntry {
-    fn into_request(self) -> tonic::Request<RaftMes> {
-        let mes = RaftMes {
-            data: serde_json::to_string(&self).expect("fail to serialize"),
-            error: "".to_string(),
-        };
-        tonic::Request::new(mes)
-    }
-}
-
-impl TryFrom<RaftMes> for LogEntry {
-    type Error = tonic::Status;
-
-    fn try_from(mes: RaftMes) -> Result<Self, Self::Error> {
-        let req: LogEntry =
-            serde_json::from_str(&mes.data).map_err(|e| tonic::Status::internal(e.to_string()))?;
-        Ok(req)
-    }
-}
